Step file separated from fixtures
