+++ tests/support/ui/content/ContentListPage.ts
@@ -6,7 +6,7 @@ import type { ContentListComponent } from './components/ContentList'
 import type { SectionType } from '@domain/section'
 import { sectionsConfig, getURLForSection } from '@domain/section'
 import { urlValidator } from '@tests/support/ui/shared/flows/urlValidator'
-import { verifyStep } from '@tests/fixtures'
+import { verifyStep } from '@tests/step'
 import { visit } from '@tests/support/ui/shared/NavigationPaths'
 import type { LocalizedPage } from '@tests/support/ui/shared/contracts/localization'
 import { mainLayout, defaultMainLayout } from '@tests/support/ui/shared/components/MainLayout'
