+++ tests/support/ui/shared/flows/urlValidator.ts
@@ -1,6 +1,5 @@
-import { verifyStep } from '@tests/fixtures'
+import { verifyStep, type Verification } from '@tests/step'
 import type { Page } from '@playwright/test'
-import type { Verification } from '@tests/fixtures'
 
 export function urlValidator(page: Page) {
   return function validateUrl(expected: string | RegExp): Verification<void> {
